Replace OpusLibrary with AudioNatives and move logic for loading
Updated dependency
